chore(IDX-3144): inline the basic_health_test This inlines the `setup()` and `test()` functions of the basic_health_test into the binary itself such that it doesn't depend on the big `//rs/tests:tests` library anymore.
